Created to a Translator class to group this code
