+++ prime-router/src/main/kotlin/Metadata.kt
@@ -409,84 +409,6 @@ class Metadata {
         }
     }
 
-    /*
-     * Mapping operations
-     */
-    fun mapByServices(input: Report): List<Report> {
-        return organizationServices.map { service -> mapByService(input, service) }
-    }
-
-    fun filterAndMapByService(input: Report): List<Pair<Report, OrganizationService>> {
-        if (input.isEmpty()) return emptyList()
-        return organizationServices.filter { service ->
-            service.topic == input.schema.topic
-        }.mapNotNull { service ->
-            val mappedReport = mapByService(input, service)
-            if (mappedReport.itemCount == 0) return@mapNotNull null
-            Pair(mappedReport, service)
-        }
-    }
-
-    private fun mapByService(input: Report, receiver: OrganizationService): Report {
-        // Filter according to receiver patterns
-        val filterAndArgs = receiver.jurisdictionalFilter.map { filterSpec ->
-            val (fnName, fnArgs) = JurisdictionalFilters.parseJurisdictionalFilter(filterSpec)
-            val filterFn = findJurisdictionalFilter(fnName)
-                ?: error("JurisdictionalFilter $fnName is not found")
-            Pair(filterFn, fnArgs)
-        }
-        val filteredReport = input.filter(filterAndArgs)
-
-        // Apply mapping to change schema
-        val toReport: Report = if (receiver.schema != filteredReport.schema.name) {
-            val toSchema = findSchema(receiver.schema)
-                ?: error("${receiver.schema} schema is missing from catalog")
-            val mapping = buildMapping(toSchema, filteredReport.schema)
-            filteredReport.applyMapping(mapping)
-        } else {
-            filteredReport
-        }
-
-        // Transform reports
-        var transformed = toReport
-        receiver.transforms.forEach { (transform, transformValue) ->
-            when (transform) {
-                "deidentify" -> if (transformValue == "true") {
-                    transformed = transformed.deidentify()
-                }
-            }
-        }
-        return transformed.copy(destination = receiver)
-    }
-
-    fun buildMapping(toSchema: Schema, fromSchema: Schema): Schema.Mapping {
-        if (toSchema.topic != fromSchema.topic) error("Trying to match schema with different topics")
-
-        val useDirectly = mutableMapOf<String, String>()
-        val useValueSet: MutableMap<String, String> = mutableMapOf()
-        val useMapper = mutableMapOf<String, Mapper>()
-        val useDefault = mutableSetOf<String>()
-        val missing = mutableSetOf<String>()
-
-        toSchema.elements.forEach { toElement ->
-            fromSchema.findElement(toElement.name)?.let { matchedElement ->
-                useDirectly[toElement.name] = matchedElement.name
-                return@forEach
-            }
-            toElement.mapper?.let {
-                val name = Mappers.parseMapperField(it).first
-                useMapper[toElement.name] = findMapper(name) ?: error("Mapper $name is not found")
-                return@forEach
-            }
-            if (toElement.required == true) {
-                missing.add(toElement.name)
-            } else {
-                useDefault.add(toElement.name)
-            }
-        }
-        return Schema.Mapping(toSchema, fromSchema, useDirectly, useValueSet, useMapper, useDefault, missing)
-    }
-
     companion object {
         const val schemaExtension = ".schema"
         const val valueSetExtension = ".valuesets"
